Set fixed snow line height for arctic maps
The snow line height can be changed in the game creation settings and
replaces the snow coverage. This should help in generating arctic maps
without relying on luck to have farms or forests correctly generated.

The snow coverage was too random, it could produce maps with snow
heights too low or too high with the exact same settings.