Set fixed snow line height for arctic maps
The snow line height can be changed in the game creation settings and
replaces the snow coverage.

The snow coverage was too random in that it could produce maps with snow
lines too low ro too high with the exact same settings. It is now
automatically calculated after the map is generated based on the snow
line height of the map.

This should hopefully help in generating arctic maps without relying on
luck to have farms or forests correctly generated.